+++ src/main/java/com/nebula/nebula_resource/helper/jwt/JwtFilter.java
@@ -1,4 +1,4 @@
-package com.nebula.nebula_resource.helper;
+package com.nebula.nebula_resource.helper.jwt;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
